Updating documentatio for https woes
I've since moved go adding https (I'll update this later)
And since doing this I've also installed nginx and using a proxy on the base server as I believe it was easier then using containers for setting this up.

In my work doing this, I severely stumbled on the absolute PAIN of switching to https.  In the end I have opted to use collabora instead of only office as I was able to get it to work.

I have provided my tips into getting this to work so hopefully you won't go through the same pain. In the future, I think I can automate these configurations using ansible so it's not a manual process.  For now I'm getting the information out here so people can fix it themselves.