Bug: low upload capacity on nextcloud
When testing uploading on nextcloud I found that I was getting 413 errors when uploading.  This was because the upload size limit was too low.  This change ensures that the maxy size is set to 0 for unlimited upload size.  You can change this to your own preferred size if you wish.